+++ crates/node/src/main.rs
@@ -7,7 +7,7 @@ mod util;
 
 use anyhow::{bail, Context};
 use clap::Parser;
-use cli::{NetworkType, RunCmd};
+use cli::RunCmd;
 use http::{HeaderName, HeaderValue};
 use mc_analytics::Analytics;
 use mc_block_import::BlockImporter;
@@ -20,6 +20,7 @@ use mc_telemetry::{SysInfo, TelemetryService};
 use mp_oracle::pragma::PragmaOracleBuilder;
 use mp_utils::service::{MadaraServiceId, ServiceMonitor};
 use service::{BlockProductionService, GatewayService, L1SyncService, L2SyncService, RpcService};
+use starknet_api::core::ChainId;
 use std::sync::Arc;
 
 const GREET_IMPL_NAME: &str = "Madara";
@@ -50,16 +51,13 @@ async fn main() -> anyhow::Result<()> {
         run_cmd.chain_config()?
     };
 
-    // Check if the devnet is running with the correct chain id.
-    if run_cmd.devnet && chain_config.chain_id != NetworkType::Devnet.chain_id() {
-        if !run_cmd.block_production_params.override_devnet_chain_id {
-            tracing::error!("You're running a devnet with the network config of {:?}. This means that devnet transactions can be replayed on the actual network. Use `--network=devnet` instead. Or if this is the expected behavior please pass `--override-devnet-chain-id`", chain_config.chain_name);
-            panic!();
-        } else {
-            // This log is immediately flooded with devnet accounts and so this can be missed.
-            // Should we add a delay here to make this clearly visisble?
-            tracing::warn!("You're running a devnet with the network config of {:?}. This means that devnet transactions can be replayed on the actual network.", run_cmd.network);
-        }
+    // Check if the devnet is running with the correct chain id. This is purely
+    // to avoid accidental setups which would allow for replay attacks. This is
+    // possible if the devnet has the same chain id as another popular chain,
+    // allowing txs which occur on it to also be replayed on that other chain.
+    if run_cmd.devnet && (chain_config.chain_id == ChainId::Mainnet || chain_config.chain_id == ChainId::Sepolia) {
+        tracing::error!("You're running a devnet with the network config of {0}. This means that devnet transactions can be replayed on the actual {0} network. Use `--network=devnet` instead.", chain_config.chain_name);
+        anyhow::bail!("Devnet")
     }
 
     let node_name = run_cmd.node_name_or_provide().await.to_string();